Linter fixes on src/components/App
